Added a src directory and moved things there for better organization.
