+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 // Dependencies.
 import { Fragment } from "react"
-import Header from "@/components/design/Header/Header"
-import HomeContent from "@/components/design/Home/HomeContent"
-import Footer from "@/components/design/Footer/Footer"
+import Header from "@/src/components/design/Header/Header"
+import HomeContent from "@/src/components/design/Home/HomeContent"
+import Footer from "@/src/components/design/Footer/Footer"
 
 // Page.
 export default function Home() {
