+++ src/components/design/About/Values.tsx
