+++ src/components/design/ContentCard.tsx
