+++ src/lib/content/getMarkdown.ts
@@ -5,7 +5,7 @@ import path from "node:path"
 // Get Markdown.
 export default function getMarkdown(file: string): string {
 	// Get the file path.
-	const filePath = path.join(process.cwd(), "data/content", `${file}`)
+	const filePath = path.join(process.cwd(), "src/data/content", `${file}`)
 
 	// Read the file.
 	const markdown = fs.readFileSync(filePath, "utf-8")
